pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "mrki"
dynamic = ["version"]
description = "A powerful workflow automation platform for modern teams"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
authors = [
    {name = "Mrki Team", email = "team@mrki.dev"},
]
keywords = [
    "workflow",
    "automation",
    "scheduler",
    "orchestration",
    "tasks",
    "pipeline",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "starlette>=0.27.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    
    # Configuration
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # Scheduling
    "apscheduler>=3.10.4",
    "croniter>=2.0.1",
    
    # HTTP Client
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    
    # Utilities
    "click>=8.1.0",
    "rich>=13.7.0",
    "jinja2>=3.1.2",
    "python-dateutil>=2.8.2",
    "tenacity>=8.2.3",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
# Database backends
database = [
    "psycopg2-binary>=2.9.9",
    "pymysql>=1.1.0",
    "aiosqlite>=0.19.0",
]

# Web interface
web = [
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
]

# Caching
cache = [
    "redis>=5.0.1",
]

# Message queue
queue = [
    "celery>=5.3.4",
    "redis>=5.0.1",
    "aioredis>=2.0.1",
]

# Monitoring
monitoring = [
    "prometheus-client>=0.19.0",
    "sentry-sdk>=1.38.0",
]

# All extras
all = [
    "mrki[database,web,cache,queue,monitoring]",
]

# Development dependencies
dev = [
    # Linting and formatting
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    
    # Testing
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-xdist>=3.5.0",
    "factory-boy>=3.3.0",
    "faker>=20.1.0",
    
    # Type stubs
    "types-pyyaml>=6.0.12",
    "types-python-dateutil>=2.8.19",
    "types-passlib>=1.7.7",
]

# Documentation dependencies
docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.14",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-git-revision-date-plugin>=0.3.2",
    "pymdown-extensions>=10.5",
]

# Test dependencies
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "factory-boy>=3.3.0",
    "faker>=20.1.0",
    "responses>=0.24.1",
    "respx>=0.20.2",
]

[project.scripts]
mrki = "mrki.cli:main"

[project.urls]
Homepage = "https://mrki.dev"
Documentation = "https://mrki.readthedocs.io"
Repository = "https://github.com/mrki/mrki"
Issues = "https://github.com/mrki/mrki/issues"
Changelog = "https://github.com/mrki/mrki/blob/main/CHANGELOG.md"
Discussions = "https://github.com/mrki/mrki/discussions"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/mrki/_version.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mrki"]

# Ruff configuration
[tool.ruff]
target-version = "py39"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # Line too long (handled by formatter)
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.isort]
known-first-party = ["mrki"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]  # Allow assert in tests

# MyPy configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=mrki",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
source = ["src/mrki"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*",
    "src/mrki/_version.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # Skip assert warnings in non-test code

# isort configuration (used by ruff)
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["mrki"]

# Black configuration (used as reference for ruff format)
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''
